# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Books API",
    "version": "1.0.0",
    "description": "A simple API for working with books"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
    "/books": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get all books",
        "responses": {
          "200": {
            "description": "A list of books",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "1"
                      },
                      "title": {
                        "type": "string",
                        "example": "The Great Gatsby"
                      },
                      "author": {
                        "type": "string",
                        "example": "F. Scott Fitzgerald"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Get a book by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The ID of the book",
            "schema": {
              "type": "string"
            },
            "example": "1"
          }
        ],
        "responses": {
          "200": {
            "description": "Book found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "1"
                    },
                    "title": {
                      "type": "string",
                      "example": "The Great Gatsby"
                    },
                    "author": {
                      "type": "string",
                      "example": "F. Scott Fitzgerald"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Book not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Book not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors": {
      "get": {
        "summary": "Get all authors",
        "description": "Returns all authors.",
        "tags": [
          "Authors"
        ],
        "responses": {
          "200": {
            "description": "A list of author objects.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "a1"
                      },
                      "name": {
                        "type": "string",
                        "example": "Jane Austen"
                      },
                      "birthYear": {
                        "type": "integer",
                        "example": 1775
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unable to retrieve authors."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create an author",
        "description": "Creates a new author using a custom id.",
        "tags": [
          "Authors"
        ],
        "requestBody": {
          "required": true,
          "description": "The author to create.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "id",
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "a4"
                  },
                  "name": {
                    "type": "string",
                    "example": "Example Author"
                  },
                  "birthYear": {
                    "type": "integer",
                    "example": 1980
                  }
                }
              },
              "example": {
                "id": "a4",
                "name": "Example Author",
                "birthYear": 1980
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created author object.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "a4"
                    },
                    "name": {
                      "type": "string",
                      "example": "Example Author"
                    },
                    "birthYear": {
                      "type": "integer",
                      "example": 1980
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "A required field is missing or the id already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "An author with this id already exists."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unable to create author."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/authors/{id}": {
      "get": {
        "summary": "Get an author by id",
        "description": "Returns one author by custom id.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom id of the author.",
            "schema": {
              "type": "string"
            },
            "example": "a1"
          }
        ],
        "responses": {
          "200": {
            "description": "The matching author object.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "a1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Jane Austen"
                    },
                    "birthYear": {
                      "type": "integer",
                      "example": 1775
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No author exists with the specified id.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unable to retrieve author."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update an author",
        "description": "Updates an existing author by custom id.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom id of the author to update.",
            "schema": {
              "type": "string"
            },
            "example": "a1"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The author fields to update.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "birthYear"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Updated Author"
                  },
                  "birthYear": {
                    "type": "integer",
                    "example": 1981
                  }
                }
              },
              "example": {
                "name": "Updated Author",
                "birthYear": 1981
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated author object.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "a1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Updated Author"
                    },
                    "birthYear": {
                      "type": "integer",
                      "example": 1981
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "A required field is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "The name and birthYear fields are required."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No author exists with the specified id.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author not found."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unable to update author."
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an author",
        "description": "Deletes an author if no books reference the author.",
        "tags": [
          "Authors"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "The custom id of the author to delete.",
            "schema": {
              "type": "string"
            },
            "example": "a1"
          }
        ],
        "responses": {
          "204": {
            "description": "Author deleted successfully. The response has no body."
          },
          "404": {
            "description": "No author exists with the specified id.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author not found."
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "The author cannot be deleted because books still reference the author.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Author cannot be deleted while books still reference the author."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "An unexpected server or database error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Unable to delete author."
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {},
  "tags": []
}